add testing to governance package, start with interactive builder unit tests
