+++ packages/cli/proposalTransactions.json
@@ -1,0 +1,1 @@
+[{"contract":"Governance","function":"setConstitution","args":["0x19F78d207493Bf6f7E8D54900d01bb387F211b28","0xa91ee0dc","900000000000000000000000"],"value":"0"}]
