+++ packages/sdk/connect/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
-  testMatch: ['<rootDir>/src/**/?(*.)+(spec|test).ts?(x)'],
+  testMatch: ['<rootDir>/src/**/?(*.)+(spec|test).ts'],
   verbose: true,
 }
