+++ packages/sdk/governance/jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  preset: 'ts-jest',
+  testMatch: ['<rootDir>/src/**/?(*.)+(test).ts'],
+  setupFilesAfterEnv: ['@celo/dev-utils/lib/matchers', '<rootDir>/jest_setup.ts'],
+  globalSetup: '<rootDir>/src/test-utils/setup.global.ts',
+  verbose: true,
+}
